Move GridHelper to own module
